+++ src/lib/promise.ts
@@ -1,7 +1,0 @@
-import PromisePolyfill from "promise-polyfill";
-
-const PromiseGlobal =
-  // eslint-disable-next-line no-undef
-  typeof Promise !== "undefined" ? Promise : PromisePolyfill;
-
-export { PromiseGlobal };